@@ -15,32 +15,33 @@
   ],
   "scripts": {
     "clean": "ts-node tests/pullImageKillOld.ts",
-    "test": "yarn clean && jest --maxConcurrency 3"
+    "test": "yarn clean &&  jest tests/masp.test.ts &&  jest tests/standard.test.ts &&  jest tests/transactions.test.ts",
+    "testaddr": "yarn clean && jest tests/masp_address.test.ts"
   },
   "dependencies": {
     "@zondax/ledger-namada": "../js",
-    "@zondax/zemu": "^0.46.0"
+    "@zondax/zemu": "^0.50.3"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.11",
+    "@types/jest": "^29.5.12",
     "@types/ledgerhq__hw-transport": "^4.21.6",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
+    "@typescript-eslint/eslint-plugin": "^8.5.0",
+    "@typescript-eslint/parser": "^8.5.0",
     "blakejs": "^1.1.1",
     "crypto-js": "4.2.0",
     "ed25519-supercop": "^2.0.1",
-    "eslint": "^8.55.0",
+    "eslint": "^9.10.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.0",
-    "eslint-plugin-jest": "^27.4.3",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-import": "^2.30.0",
+    "eslint-plugin-jest": "^28.8.3",
+    "eslint-plugin-prettier": "^5.2.1",
     "jest": "29.7.0",
-    "js-sha256": "0.10.1",
+    "js-sha256": "0.11.0",
     "jssha": "^3.3.1",
     "leb128": "^0.0.5",
-    "prettier": "^3.1.1",
-    "ts-jest": "^29.1.1",
+    "prettier": "^3.3.3",
+    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.6.2"
   }
 }
