@@ -17,105 +17,87 @@
 import Zemu from '@zondax/zemu'
 import { NamadaApp, Signature } from '@zondax/ledger-namada'
 import { models, hdpath, defaultOptions } from './common'
-
-const sha256 = require('js-sha256')
-const leb = require('leb128')
+import { hashSignatureSec } from './utils'
 
 // @ts-ignore
 import ed25519 from 'ed25519-supercop'
 
-
-function hashSignatureSec(pubkeys: Buffer[], salt: Buffer, hashes: { [index: number]: Buffer }, indices: Buffer, signature: Buffer | null, prefix: Uint8Array | null) {
-  let hash = sha256.create();
-  if (prefix != null) {
-    hash.update(prefix);
-  }
-
-  hash.update(new Uint8Array([indices.length, 0, 0, 0]));
-  for (let i = 0; i < (indices.length); i ++) {
-    hash.update(Buffer.from(hashes[indices[i]]));
-  }
-
-  // Signer::PubKeys
-  hash.update(new Uint8Array([0x01]));
-
-  //Pubkeys
-  hash.update(new Uint8Array([pubkeys.length, 0, 0, 0]));
-  for (let i = 0; i < (pubkeys.length); i ++) {
-    hash.update(Buffer.from(pubkeys[i]));
-  }
-
-  if(signature != null) {
-    // u32 representing length
-    hash.update(new Uint8Array([1, 0, 0, 0]));
-    // u8 representing key
-    hash.update(new Uint8Array([0x00]));
-    // common::Signature
-    hash.update(signature);
-  } else {
-    // u32 representing length
-    hash.update(new Uint8Array([0, 0, 0, 0]));
-  }
-
-  return Buffer.from(hash.array());
-}
-
 const TEST_DATA = [
   {
     name: 'bond',
-    blob: Buffer.from('1d0000006c6f63616c6e65742e6664633665356661643365356535326433662d300023000000323032332d31312d31365431343a33313a31322e3030383437393437392b30303a303029e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be0101000000000000000000000000000000000000000000000000000000000000000032fdd4e57f56519541491312d4e9089032244eca0048998ffa0340c473b72dad3604abd76581e71e4a334d0708ef754a0adcec66d80300000000000000a861000000000000000200000002b3078bd88b010000007c7a739c83e943d4a56a0fd4e4c52a9edc0d66d9105324bcc909619857a6683b010c00000074785f626f6e642e7761736d00b3078bd88b0100004b00000000f2d1fbf5a690f8ab12cfa6166425bec4d7569bb400e9a435000000000000000000000000000000000000000000000000000000000100ba4c9645a23343896227110a902af84e7b4a4bb3', 'hex'),
+    blob: Buffer.from(
+      '0d000000466a5f5f5f445f6158595f5f4b0123000000323836322d30382d32345430323a35303a31352e3333373034383037392b30303a303023000000353631332d30342d32355430333a33323a35372e3535313732383231302b30303a3030010000004e7dabcefd260e5ee60fa452f53049bed3ab350e0cd9c7a82d6339f9b07ee94e06f2a8a5d19e5100c7b9343908f9793a1ce7a37aa89e6813147cf6b62da108f600000000000000000000000000000000000000000000000000000000000000000001b895ea1c1de9a22f00000000000000000000000000000000000000000000000055007fb7423cd0f49513025620f11fde3a29381540620028aeac349e51d066d1850fe25da3d3d323f587a89aeeef07d91456cc044335415a320b252c097dad03000000009e63e0a48f0100003600000001235d80cf0724fa7a6015b2e11e5cbfb511c82218483e01b8f2596ff100000000000000000000000000000000000000000000000000029e63e0a48f0100000089147c3a90267516247222324800c63d9a4a4db4e5e148f30f92ac5fe9954649010c00000074785f626f6e642e7761736d0303000000425aeef657034b6ff6e7e7329bd2ebe4ab0630b687b574f93f3451627f93843806f2a8a5d19e5100c7b9343908f9793a1ce7a37aa89e6813147cf6b62da108f64e7dabcefd260e5ee60fa452f53049bed3ab350e0cd9c7a82d6339f9b07ee94e0102000000003698bfa03124a04aa71206732efb989983d0a7c22ecede0bb75d45865588484000cf2a13ff9d67a2f407d68149291dd9012ede5436c312612aae6593009f9e5847020000000000d9057e8be1ede8c772d76a6df6dcb35a6e2c4ec212528dcfb5bea55055c4c0eb17ae3e75bcf92333f986773d3e44b72b64e3f22d35f995a22de33ec18ea63a000100e143a9cccefceb7563c784fb7acfd9979c7b5c3480d9123a8f12925bf520084ee521ed17f5da5217a092ef50cf888fdf587a1b2406106585b582a7f2f0e54200',
+      'hex',
+    ),
     sectionHashes: {
-      0: Buffer.from('5b693f86a6a8053b79effacd031e2367a1d35cc64988795768920b2965013742', 'hex'),
-      1: Buffer.from('29e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58', 'hex'),
-      2: Buffer.from('f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be', 'hex'),
-      0xff: Buffer.from('c7fec5279e22792a9cad6346f8933c1b2249043e1a03c835030d4e71dfbac3e0', 'hex'),
+      0: Buffer.from('425aeef657034b6ff6e7e7329bd2ebe4ab0630b687b574f93f3451627f938438', 'hex'),
+      1: Buffer.from('06f2a8a5d19e5100c7b9343908f9793a1ce7a37aa89e6813147cf6b62da108f6', 'hex'),
+      2: Buffer.from('4e7dabcefd260e5ee60fa452f53049bed3ab350e0cd9c7a82d6339f9b07ee94e', 'hex'),
+      3: Buffer.from('54c94e92ac1955bd02c43805c47cd997fac45071b6a2407bd43f2ac9fb06cef7', 'hex'),
+      0xff: Buffer.from('be87d012422f1a676686f1c718af1b9b77dbf47875aed905bcd326fc13cd8f51', 'hex'),
     } as { [index: number]: Buffer },
   },
   {
     name: 'init_proposal',
-    blob: Buffer.from('1d0000006c6f63616c6e65742e6664633665356661643365356535326433662d300023000000323032332d31312d31365431343a33313a32382e3130313832393531382b30303a3030036244dbf2c2d2fc6900e7c46a0d81e75a6e137c4900ee614cc8b1551543ce5d9e467e3e8db73280a50aac95b59e54ece4f670624e1a46388dd0c458e78076420101000000000000000000000000000000000000000000000000000000000000000032fdd4e57f56519541491312d4e9089032244eca009fd0df101ba3e91d24f555893ae1bf0b271bd98363d6d4c659876731fe4a75170500000000000000a86100000000000000030000000194468bd88b010000007e68fb834a7772c82a312c4e6e519d97282cce39507950c35fe89f1c347a4a2e000294468bd88b01000000d81ff4dba9b1316576360dc94fcd611f9b228b040eaefb5157819d692eb4ef50011500000074785f696e69745f70726f706f73616c2e7761736d0094468bd88b0100005000000000061a4caa2da23123adeb4ae80ba0dc5239bff64baa4aaf5f385dc6877ae22e2600282bcb6a66c770f23e7b2f068ae0ba025eccb2ec00000c0000000000000018000000000000001e00000000000000', 'hex'),
+    blob: Buffer.from(
+      '150000004f5f5f4a5f3573474e733266516f743772585a455f0123000000353738382d31322d32355430343a31353a35332e3532353833343830332b30303a303023000000303234352d31302d33315431393a31393a32322e3634353135313830312b30303a303001000000e5ee4d20d4c136a3183ba769146081a021d212043b31999d198325f6f87798e77272069b87770918731bbb6909399fc04aa831d008507b74efd1c8f0e197ba0a000000000000000000000000000000000000000000000000000000000000000000015d07c053bbfd22ed000000000000000000000000000000000000000000000000c400853a3b123a7f71e8692991545cc51774d30f6502009b6866d17c4310ece3424278ea9576e3fed7ff34cddfa6c22d3c8844705daf336b97dba61d9b6c99050000000135c4753fa7914d2d01600200004faa8bf4b560fc483479119a7e0fe7cb675946759997dd584ed75035ff7ef57a54e4a7d9da4c3261edd860be891d9e7adbe618a2a10ff06954a544639e3ed9c3e0bc9f9774950881f12268723803f86591adb8bc139d205075b841f06f8187d74adaa8ece22883faf21b1f8f9661e0d360b9a50358f382d1bc1deb9fd622ad9ae730f3520e2942aed0c16916a4e73b4129e2b73f4b97a1d6aaf9b24df93f78546d05419164e8bc8e372ed64fa0dee900d5f346e071857801c40cfb28aad31bbb04ad7b26c75065581281ed2cf2b904678b24af76f92368331e4f399e08d6726154f7739b3ba9a80cdf2c68c8097d64d8f216ddbfac5028715a72235e56b5bc4e5137581c78d014a833ff7721e6dbf36c599efe7815200cd094789506e8f4f3dde091ec2ad9bd304b993910c0a0ba978e954cd9b443ac98b36373f507d87ef49021fdc66a2e8dfc7e3c4355b1ef4cf1888ad317cbe45899ce5f81ddf7b9490a16bcfbeaa6c1068dae27c01b455e657d19bde83199932e1d8c437fa3d04d7006e517430890dbac13791c8980a8b4b0f65c0bc0617cc092ae8e16b8d032cde9091486c96bff374b576628c70e168ac41bf912889d95337eb734ddb0470190465ac7c3af0d3453bba369d4980fc1e355e2147785f9c26bd65f7e02e6cb7ecf9e6f1f8f1e39e05ad9aea347dfea62128bdc51ed65cea0d1469c1b0da8c3ad6ff211d0434221349cd1d1255e7fa579d43a0c4f00eccb3685b271db2813fcb709b0ed9ad35bdea534176d3329d0f7968dd38871b2a9c4e8646914c1fdf60497b118fc7de521c32fc27764af13231ca46c6f2b0002bdd36dfad2ca1f676fa564d841adeb0114000000386b5f5f73305f4b6843654d39524e436f787a5f00c963e0a48f01000068000000179a46bb0e8f601b005456eceec5ebc152adc876139ef61e4ab4646007896b95018323b22fc05966adb05f065d12933d7313e5301b020100000000013a17a996503b297cf6bde37328e734ce33fcc3363f247caa524a48c53756ab906aa9b76311451c3f4af80eb902c963e0a48f01000000b7cbc4cefa2a5a1cea1e38ca1226b45d199f8784bf49dabaa7a300b69aa79d63011500000074785f696e69745f70726f706f73616c2e7761736d0304000000e5972e29aec40eeca07254e4f1c09d6d0bc420321cb71b17058976480655eaff179a46bb0e8f601b005456eceec5ebc152adc876139ef61e4ab4646007896b957272069b87770918731bbb6909399fc04aa831d008507b74efd1c8f0e197ba0ae5ee4d20d4c136a3183ba769146081a021d212043b31999d198325f6f87798e7010200000000d9eb70cb1d1e74ca26cfaf3d3767f276eadb4a06183ff3e078492af5b3a6f06900911f7c56718800d8b44f19de785776d643acbd77ea32e7f96a387a1bde59b75c020000000000e7008e509be0e290e57f224a59cc05c56c5b68009869ba26f23e7d9f49ad541abec1cd342164cfcd63d62080a7a7203748d6b5bf7e983258f47103493ad14b0c0100803de1c3ac908d8031c710dbc18f7f8932cc3fd115c00e681e0fb7b56f4a123a064316df6c64f7ab0618d720cfbfdaacbe30642c3d9cfb98c1fe0d81c75174050304000000e5972e29aec40eeca07254e4f1c09d6d0bc420321cb71b17058976480655eaff179a46bb0e8f601b005456eceec5ebc152adc876139ef61e4ab4646007896b957272069b87770918731bbb6909399fc04aa831d008507b74efd1c8f0e197ba0ae5ee4d20d4c136a3183ba769146081a021d212043b31999d198325f6f87798e70101000000000f1e7c670cb1f32cbf4ea7417bfea80fd533045e010bde397ea125c13af2e546010000000000fe133b80eb9ef712ac2dfdd015e97a52c9a275183e7918fefc6b2bf776a682ec026a7b36fc5ff79ed31644c039a6d5311300b7e013b4a850bc661ca8e545c70f',
+      'hex',
+    ),
     sectionHashes: {
-      0: Buffer.from('e86108905c4ba210c2c20289369b9cbcad3dfed25c601b9b6673ddef5573e3d4', 'hex'),
-      1: Buffer.from('061a4caa2da23123adeb4ae80ba0dc5239bff64baa4aaf5f385dc6877ae22e26', 'hex'),
-      2: Buffer.from('036244dbf2c2d2fc6900e7c46a0d81e75a6e137c4900ee614cc8b1551543ce5d', 'hex'),
-      3: Buffer.from('9e467e3e8db73280a50aac95b59e54ece4f670624e1a46388dd0c458e7807642', 'hex'),
-      0xff: Buffer.from('039a4d90beb2ead23ef48cecc0039104f865dd3c6c89ee12b90defbe8205c319', 'hex'),
+      0: Buffer.from('e5972e29aec40eeca07254e4f1c09d6d0bc420321cb71b17058976480655eaff', 'hex'),
+      1: Buffer.from('179a46bb0e8f601b005456eceec5ebc152adc876139ef61e4ab4646007896b95', 'hex'),
+      2: Buffer.from('7272069b87770918731bbb6909399fc04aa831d008507b74efd1c8f0e197ba0a', 'hex'),
+      3: Buffer.from('e5ee4d20d4c136a3183ba769146081a021d212043b31999d198325f6f87798e7', 'hex'),
+      4: Buffer.from('bca7a9f1f7aebadbcf64578d1f48774b814598740b9c610d69a6b205881d8c78', 'hex'),
+      5: Buffer.from('b2842bb77b42908f56f4a061dd56a3e77922f9002c25370c6468fdfefdf4a03f', 'hex'),
+      0xff: Buffer.from('f05532c2dd0bf7554706ef2a684b96521ce1ab04a39abc3ce16d4423d7146a1f', 'hex'),
     } as { [index: number]: Buffer },
   },
   {
     name: 'update_vp',
-    blob: Buffer.from('1d0000006c6f63616c6e65742e6664633665356661643365356535326433662d300023000000323032332d31312d31365431343a33333a30352e3139313137373835302b30303a30305df9287d1f3b893eef4643c47a4d40a8838c9fc6a39bbda275e20c308f6bb0a290f781fd267026dfd1335d092f0a46d9cf00b8816c5997e55eaf88e0e5cbf3dc0101000000000000000000000000000000000000000000000000000000000000000032fdd4e57f56519541491312d4e9089032244eca009fd0df101ba3e91d24f555893ae1bf0b271bd98363d6d4c659876731fe4a75171400000000000000a861000000000000000500000001d2c18cd88b0100000053e9023f2edc55e90bdcf6024760140a91a7904cb4a529cb1707ea3d34ee685a010c00000076705f757365722e7761736d02d3c18cd88b0100000049a9b2023b4858255845bfa3949e314f241858949214950471b0b598febd7248011600000074785f7570646174655f6163636f756e742e7761736d00d3c18cd88b0100009f00000000ba4c9645a23343896227110a902af84e7b4a4bb301a911dbc336d7feab85e374e331542d94155fe25219bad6d7f497a2e5d9d6a41703000000009fd0df101ba3e91d24f555893ae1bf0b271bd98363d6d4c659876731fe4a75170048998ffa0340c473b72dad3604abd76581e71e4a334d0708ef754a0adcec66d80075c251407a7ed91b8114f90b6f454db9071a290458e2db36106b99daf4bf83b101020301000000ae19fb14011c761f61ea480a4d9fffd1abb049be4715f5344d6f60093470c0800000ba4c9645a23343896227110a902af84e7b4a4bb30100000000000d1cfc74bd04c329f14c466b7b17ac72b00818db85bb1069415a898bfa95cd21ae416cfc38e190ff2a631f1f9dd700556d5ad25eef92a5f5dffa5af3a723390c03050000009ac0dfc46072a5736aa2257015d54ecb5490f6811bf9ffee38662c38cecb1ea8a911dbc336d7feab85e374e331542d94155fe25219bad6d7f497a2e5d9d6a4175df9287d1f3b893eef4643c47a4d40a8838c9fc6a39bbda275e20c308f6bb0a290f781fd267026dfd1335d092f0a46d9cf00b8816c5997e55eaf88e0e5cbf3dc66f097755345c31d1ccf8b764bbbd8ff8b8496f7ac88a32f4967b98138b642aa0101000000009fd0df101ba3e91d24f555893ae1bf0b271bd98363d6d4c659876731fe4a7517010000000000e7c0e073a254bccce92bb5d6b35da86f6ee1b300fa6d3f54d1414a06745a895d32d5b86fe029e0ab456d2ade56f79d066f3e9c6dd80088c9fdc6e00a4a472f0a', 'hex'),
+    blob: Buffer.from(
+      '16000000424f394e6c565642433755487a675276756f66625f5f0123000000323236302d31322d31345431343a30353a30312e3132383133343031342b30303a303023000000363638342d30352d31375430313a32313a34312e3133323932303538372b30303a30300100000062c02f7fab4fb0e846e62b99550b1f4c45218d9b8c194483a7da5bd0787f6dc3cfbfe5addf8230284ab2f4b1121634c5a8a67513375f7322018e1601578bd29100000000000000000000000000000000000000000000000000000000000000000001b7d061a14fa029700000000000000000000000000000000000000000000000001600f0a7fb550bf7408a16896720bdfc2700009833bd01021419f36287b7c0795f26e5c92b020bb89dcc9ed2d67bff237dab3bc310c8d8255fd55c7546519df80400000001b9d20c1e1933ec29019f03000074b3e740478d80e44b41e3afa1514db55b8e98359a224d06bc6224280640285865cd6d33e64725bcf0bdfefc1d979f381161ec7ba44a211beefc6de325706698db7d4e4049432f2fadb7b93c8baa23c8540e2095ced6e0b86d405892b2a3617f31e04a6256542fd5022befaed6949ea4784d775790ca77a6e7e8c8728f6d2a71d0444e8489e73e0b73cf25a296caed20ab9d7ac357573e8ca304027a353476b5efbe115e002fda981b12085a746105200afc3ab64e51fad3a32efbbc2f3eacdf240b49b8148c6aa364ad284233170157034f7cbf156c98a849144c5cf5382baeb23884a6fee204b51d79f5888a518385567c4715cb10fad2256895a5b00c843fe0bab315b706c1f720592cdd70d83255fb508811043dc47954fc87b679a4c15ce6c00e7de5abfde440d451e8016b32b379227c63d9ec29e978a58303cd31ec461be28e79125d9e186f0ffcd693cf680f7e02938beac4adb51fda7a8591f7d0ebe8322570301137207617c7d2409e7cf685b1a40d330fd292133ba10fbe107f89cdc9927755d3adad2214992148b30d9f47d57dfa1f35575479d5e837750efa21b098ae819992bc840e7eedf03e9cd04882505c1871b1e8ed0fa233b1a8cdbeb2c24199028982d184daaf932f5de68e00e47afedda194f655f40d341ebb3d66b9815092b1353ea2098a132ccef89ab9e1c1df7121f8a94f9e224e333967f4bd0f8425fa734d0416646ea9025aae32b0e1d6b78f73e95f4e43b27b31bd6620e133fe0d3024ae29121007da1f7ca43d0ea80c8e5458a9d983a0d358dab59de3bbf896c0c9ac031655e7bcc7e569bd29e12eb4a6b152254e28b6a3c99cd6f059fb6cb2d9130ef33d57e7385d8e0a319a9b559dbd45447d0f8876233c944fa8ab5a83fb5cb6ff2aaf699cc1598dc33601a65c32f2584816aef4e78592b901136bc20bbffca17f564c780c2e7a20d962759000204567987de5f6074fc7da95cab5ae3860edfb85ee376b321e054b16df08fc0508626fd252757a78d97ee4ccf1100727707148b059087f1ae32f1d48a50c8604f73003beec0c0b1d09152cf5f2aa6454fc488b9c16fb0c84a980292af6dd38fbc1200abaa8946fbeeaebe5b0f825496b8f85404a1b9ba43ca3fe8a2054d8494aa61a88f59709735f2f760246b2d14b1be3ea420396d1bb3f72e70ee121cf0c0e237b8bc7a929962bddc71837ee006124985f579e442809338f9bba759ca6202d8b4928e6493686cbd23cfda64bb489b90bd71391d2091506fc528814fd52373a04e5c35b82e5c680203391d2bff5820000bb65e0a48f0100000301000001d2eaaeb78e51479bad7228ac23fbcc1726218d9a013afcac29972d3b65ca62fe6c59d7a8f804f4053d2e880673db00eaec0373c456060000000021e47c68dcf4579dbe327751cc69c9f42c4519f47aa43c289fa01aa91f89e96c006296b1ec7897c15a58800c9e8fffd38400d3a4c9b8d93c48b53d23273e7fdc0e000c81fa92e02bb942cc7e2b00bd9f2b5daa0256b143c43dc98fe3a44deeca1fcd0029db8a369ea03ea6a1ef7033e74095996db326384d6ad5e0f36720a7d37262230102328b1c1001d997b6c44fabb7b74a176be43d2583f86ec5e1483bc5719384b8b4010282048df0c40b3457aea04bc9fac81d51640fd197022d11c999942a768bc7965d0002bb65e0a48f01000000f6560569dc543cbd763d583e74bc2a06157c3cb2560e0bb4551ab8b30c9fe79a011600000074785f7570646174655f6163636f756e742e7761736d0304000000bf8d079ebce675e54b3ca4d7619dba09a9fade9dce2ef3ad2890a1424b6533f53afcac29972d3b65ca62fe6c59d7a8f804f4053d2e880673db00eaec0373c456cfbfe5addf8230284ab2f4b1121634c5a8a67513375f7322018e1601578bd29162c02f7fab4fb0e846e62b99550b1f4c45218d9b8c194483a7da5bd0787f6dc3010200000000b8c2b9f5543592a2fa3325868661aa0a2ea04609b615249a109321a14ae2eae000d023db0cad154b5ad0914b72ee07ec620f55ade8794ce381965d3e23f9d47bca0200000000008539c01b2487882f19e295e43e15bd97ce6ce0d573ab162bc387bbebcfa4199162a4c399e3d0ea9a63fc6d2247125ab30863db7a2b6f4ce2185181b81dd2140201009ef99f8a4cf18ed0d37b297dd8c8bfbe5fb5529fe4eb09c00d3e140c2ab03958264fae0d85d9889467de57a9e4ed29dbaecea5daa9b5c1d91f71892c2660300a',
+      'hex',
+    ),
     sectionHashes: {
-      0: Buffer.from('9ac0dfc46072a5736aa2257015d54ecb5490f6811bf9ffee38662c38cecb1ea8', 'hex'),
-      1: Buffer.from('a911dbc336d7feab85e374e331542d94155fe25219bad6d7f497a2e5d9d6a417', 'hex'),
-      2: Buffer.from('5df9287d1f3b893eef4643c47a4d40a8838c9fc6a39bbda275e20c308f6bb0a2', 'hex'),
-      3: Buffer.from('90f781fd267026dfd1335d092f0a46d9cf00b8816c5997e55eaf88e0e5cbf3dc', 'hex'),
-      4: Buffer.from('66f097755345c31d1ccf8b764bbbd8ff8b8496f7ac88a32f4967b98138b642aa', 'hex'),
-      5: Buffer.from('521915f7e75cbde1d94d178a83f6c4ca51bc77640fcd3aaeb28b89ac16620d58', 'hex'),
-      0xff: Buffer.from('ae19fb14011c761f61ea480a4d9fffd1abb049be4715f5344d6f60093470c080', 'hex'),
+      0: Buffer.from('bf8d079ebce675e54b3ca4d7619dba09a9fade9dce2ef3ad2890a1424b6533f5', 'hex'),
+      1: Buffer.from('3afcac29972d3b65ca62fe6c59d7a8f804f4053d2e880673db00eaec0373c456', 'hex'),
+      2: Buffer.from('cfbfe5addf8230284ab2f4b1121634c5a8a67513375f7322018e1601578bd291', 'hex'),
+      3: Buffer.from('62c02f7fab4fb0e846e62b99550b1f4c45218d9b8c194483a7da5bd0787f6dc3', 'hex'),
+      4: Buffer.from('b6b457fef2b4b40ae1305146d42fa5f37d9c67cbe8457a6900d0122e9f66416b', 'hex'),
+      0xff: Buffer.from('7a6a22cfca3ba5d9edc19336327ec772999b6debffc39fa6de70281190fdc0e8', 'hex'),
     } as { [index: number]: Buffer },
   },
   {
     name: 'multisig_pubkeys',
-    blob: Buffer.from('1d0000006c6f63616c6e65742e6664633665356661643365356535326433662d300023000000323032332d31312d31365431343a33313a31322e3030383437393437392b30303a303029e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be0101000000000000000000000000000000000000000000000000000000000000000032fdd4e57f56519541491312d4e9089032244eca0048998ffa0340c473b72dad3604abd76581e71e4a334d0708ef754a0adcec66d80300000000000000a861000000000000000400000002b3078bd88b010000007c7a739c83e943d4a56a0fd4e4c52a9edc0d66d9105324bcc909619857a6683b010c00000074785f626f6e642e7761736d00b3078bd88b0100004b00000000f2d1fbf5a690f8ab12cfa6166425bec4d7569bb400e9a435000000000000000000000000000000000000000000000000000000000100ba4c9645a23343896227110a902af84e7b4a4bb30301000000c7fec5279e22792a9cad6346f8933c1b2249043e1a03c835030d4e71dfbac3e00000ba4c9645a23343896227110a902af84e7b4a4bb301000000000087d6e5a4617cce4c93120504a5f5db8c9ce1af0416e260c3fbe9066df3f3fdb2abfda0cac21b97b3e89b3c29013db345bd22548e8baf2df4e682bb4e1a041f0f03040000005b693f86a6a8053b79effacd031e2367a1d35cc64988795768920b296501374229e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897bed4bfd3e247c0ef6e2ab23983a793412fd94a78d9a08efaa94a3d6a977e3c601c01010000000048998ffa0340c473b72dad3604abd76581e71e4a334d0708ef754a0adcec66d8010000000000cfcc82f327627fed72368dd168663db755478675d812365b9c8b92c36acaaebf1fe9a0494aaf9e675d4b4f041ffebc5234d9da012721b1bd5d1bbc819ed56f04', 'hex'),
+    blob: Buffer.from(
+      '0c0000006179456a5f4a45595f7a336b0023000000343032312d30342d30365430313a31323a31372e3636383136343031312b30303a303001000000c29b111bd28d49dfdffe0c61bfe335ff3c5023309acc038c4b4280b2142115bfd331f90d44721c7d313298a28a45797d725cce9ec6015381ed88ccb4c680da79b2f773b1a9d7e6c4b647d1928009fd50b74c64e1c2b55e3c252cffaf4900be4800017b0873efb9ad927a000000000000000000000000000000000000000000000000370048c2e4e94bfb268e06bc9c4600e04a9548703ce1001a2d2e49d8fc756ce6844024ce1362c02b6c6be36f061a0ae630725d90e55d623856f1119d11c3ed05000000007a63e0a48f010000a80000000090509dbf7e788b4a779b984e5cf8b92d0286f84b75c8ed255064d28abf2bf6d92dd411c9de261178006a78aeac6ba3c57eda7b7dcf52e8418835f3d8b96004392d0acb2b760000000000000000000000000000000000000000000000006c4f8b1181132f230000000000000000000000000000000000000000000000000177ff497aa4d0e356e749080eef62d1896ec95cd400cbbb31f9b2ba56c76c1c45df1c527dad80a99b2e027a63e0a48f01000000288cc64131abdb143deefd074b7b7a586103a1c516093e8aacbbe365da84cb06011300000074785f6272696467655f706f6f6c2e7761736d015a0134744089541b006dd6d88edd63e86070f6e6fed0c66d184007cc5ffd757ee08837fe2bb8831e27014b0300005f5f324a425f305f4b775f4a5f6c3338413231795f365032345756395f434a784f794f5733556e78325f4e30305f596d734155545f3939504f6a644d73315f686c347038745f6a325f4d5f77415f3436755f505f5f5f6f396a384c4f786153723647465f44346132485f586978474b67387a6c4d5a39695a36774d5f786f4e715f5f345656335f79745f5165535f70635f51775f33325f5a33473371326b767a643839554232766e594f49764e5f5f304930335f6f304a305f6f396558375f5f7931596c795f534c747759457041344b6945314b516b5f775f5f33317159356d4b30705f76415f353041435f35557558777153624a584c50315f3568325f795f5f365f595f684e5950765f305a4239673230745f4f5f33686b666a52474a4f6c3067703267645f30395f58365f4e495f4f6d7335385f544d5837366c695f355f516c46755136445f5f736862545a5357353373305f6d674447597242557856395f5076365f76666c547a5f7a5f39686a4b396d574d4b564c5f4b6b6932365f6f51394a5a356878656d506146307379495f59686f57754738304e775f6c45556a42496b5f386b546b5f46763053614b5a3634305f5f6f394a5f6931395562394f5a5f305f6c6c41587774706d636f4a5f66393557785f6e715151624435466b4d7075525835684e375f6a305f3079356f395f5f7a5f506243766b69466344317938314c595f4d483550706f5f46365f3766365463525f566b5750737477505a345f4a76535f334d5f5a544a6e42636a33315f3841425535555f313964365157415f4848507a43684f31765f426354785f386769744f5f6836505137394d75376341557148695f36444d4a70566d7633315f4b375842686b53435f415f5f4455674b5f4c5f5565545f30475f5f5a57673335305f62337838535f5539344158665f6b6b4e4d56375f5f4c594e556d6671374b6f375f4345414158775f426c5f685f47455f6e6f763779315f6b354e6d5f785f3647315f705647315a5f67375f725f775f78796b4431365a386642306a306e5f5631756939524b31665f6a5f4c4c735f5f4b5f5f745f4778485f3062733369565f6d345f46303950714a5f5f4b7230415f5f395541465a345a3969634f726c3161395f355f5f5f6f5f64655f5f67357435745f65567231565a45030400000060a5a573a3ce1022133e0a117d56d0ec349439fd47784f9605b545e455a52029d331f90d44721c7d313298a28a45797d725cce9ec6015381ed88ccb4c680da79c29b111bd28d49dfdffe0c61bfe335ff3c5023309acc038c4b4280b2142115bfb2f773b1a9d7e6c4b647d1928009fd50b74c64e1c2b55e3c252cffaf4900be4800009a9035ccf0e0ba2b94d5b9a1b5b4820dfb50cfef0200000027008acb6d00f04e6f66021aac3b2e22c79fc8c78c60ded203e61971c3b7c84dd2e1a2c94e239890fd54ea552268aea7a74de905ec8a1b3282189ada8277803f180b7d0039b5f0b787d4cc4e60ce2a9800c07e68fbf66f7c9515b2bc3b4286c6bb29d8d88d1d8653cd49c84c7cb2da9ff9525cf02d248be8ec67376fb9e559fea3fed40a030400000060a5a573a3ce1022133e0a117d56d0ec349439fd47784f9605b545e455a52029d331f90d44721c7d313298a28a45797d725cce9ec6015381ed88ccb4c680da79c29b111bd28d49dfdffe0c61bfe335ff3c5023309acc038c4b4280b2142115bfb2f773b1a9d7e6c4b647d1928009fd50b74c64e1c2b55e3c252cffaf4900be48010200000000a4f79d3136d5704cffe650c44ebbd3b00204c8ee597b075ea8fcf914e4758683005a1b6b413136acd9a291bab5c96dd02572cc4180a7cfc40870e257c524476165020000000000be31eca5557263f3c2baa361914dd1bce4a7d03c999d7a1b860bdde41be4bacd60932f66a622d4eecf2d2337ede444edbf50abfc39885e1b4b6ff4deb2fb2a07010067956f312007a2ec31ee68dc850d5b9a441a52a45ed9cd59a1323eddcc6a316f8b6ec0c8310cca84d8785e3c9ed61ce24d5c1a48d6b340250030dc56ca927e05',
+      'hex',
+    ),
     sectionHashes: {
-      0: Buffer.from('5b693f86a6a8053b79effacd031e2367a1d35cc64988795768920b2965013742', 'hex'),
-      1: Buffer.from('29e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58', 'hex'),
-      2: Buffer.from('f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be', 'hex'),
-      3: Buffer.from('d4bfd3e247c0ef6e2ab23983a793412fd94a78d9a08efaa94a3d6a977e3c601c', 'hex'),
-      4: Buffer.from('ea7dd39da3e99c29ee2c51d7bc7cd14754010fa16531ff807988c5018fcccea4', 'hex'),
-      0xff: Buffer.from('c7fec5279e22792a9cad6346f8933c1b2249043e1a03c835030d4e71dfbac3e0', 'hex'),
+      0: Buffer.from('60a5a573a3ce1022133e0a117d56d0ec349439fd47784f9605b545e455a52029', 'hex'),
+      1: Buffer.from('d331f90d44721c7d313298a28a45797d725cce9ec6015381ed88ccb4c680da79', 'hex'),
+      2: Buffer.from('c29b111bd28d49dfdffe0c61bfe335ff3c5023309acc038c4b4280b2142115bf', 'hex'),
+      3: Buffer.from('b2f773b1a9d7e6c4b647d1928009fd50b74c64e1c2b55e3c252cffaf4900be48', 'hex'),
+      4: Buffer.from('baa3fe6346b5ae0be3e3db8d8d88026dc0b71d6c9a43d98db0d5115e9f3ebfb7', 'hex'),
+      5: Buffer.from('01ee3dc3ad219ce917d1cebb4f9fc6303f618d36bc5f9e49b3c5adf81a965195', 'hex'),
+      0xff: Buffer.from('09fb3fb4eafb7f20be93eb71767959341b8c69c914d96eb00cec07d62dd34b77', 'hex'),
     } as { [index: number]: Buffer },
   },
   {
     name: 'multisig_address',
-    blob: Buffer.from('1d0000006c6f63616c6e65742e6664633665356661643365356535326433662d300023000000323032332d31312d31365431343a33313a31322e3030383437393437392b30303a303029e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be0101000000000000000000000000000000000000000000000000000000000000000032fdd4e57f56519541491312d4e9089032244eca0048998ffa0340c473b72dad3604abd76581e71e4a334d0708ef754a0adcec66d80300000000000000a861000000000000000300000002b3078bd88b010000007c7a739c83e943d4a56a0fd4e4c52a9edc0d66d9105324bcc909619857a6683b010c00000074785f626f6e642e7761736d00b3078bd88b0100004b00000000f2d1fbf5a690f8ab12cfa6166425bec4d7569bb400e9a435000000000000000000000000000000000000000000000000000000000100ba4c9645a23343896227110a902af84e7b4a4bb30301000000c7fec5279e22792a9cad6346f8933c1b2249043e1a03c835030d4e71dfbac3e00001ed03655318474529449ed5f0bd81d2cbfa41d57a010000000000682203cfc3d10fd4bd2fbf57181012ca8f113e0b11ae6cd50621a2acf34e83b755fed6eeeb9f6f1c4f55765e2999ce8bd505fb48845e5f810ad57673fcb38e0d', 'hex'),
+    blob: Buffer.from(
+      '160000006b755f346d316d4c42397748374b65323871675f59500123000000323536312d31312d32325430363a32363a34362e3737343333303530342b30303a303023000000383934322d30382d32315430313a32363a31382e3037363937323734392b30303a3030010000003d907d242c8834e750cdf0f7f12797f1d8aced6448b14b7e6e491256539695ef79a6319f225a0d92e01f15b5e673d787ab05194d7c39e126e2eafbc2fc1ccc6b57a99887cdd729ad768b12a694a37a0e524ec6d68b4a5bbd3c90956bb2d536060001298ba55b51d1a3f600000000000000000000000000000000000000000000000071001b9d135a468a70fa3acb3bf8837232fd244658bf010211dd5fb93c373f45bd5ee6fc50edac32ce262dbc8e5701ea839c625cfa72096fbae2e82f3950358705000000009463e0a48f010000220000000103d72eb5b02f252878ea27fde9eea31349a85081d71aadd51ed10a0fc50b9e690b029463e0a48f0100000092e1536c998dcaf5d12147de83fbf68158a3731dbe5b35a59a5eb2a180e38b2e011100000074785f72657665616c5f706b2e7761736d01789e92adbec8af47019a0100007a303833314932315f5f35374f41625f6438364735485f3242634a30355930785f7466445a5f5175325536665f47577435323151305533423037454643715f327730625f465f3241385f5f45325f3152324a784249685f635f305f68515f3449345f5f676943506272394c32375f5f58693374476d345944724f6f32366230307038323072675f765f45375246755a6c5f4b563542494f54755f5f425f74765f5f367a5f384d375f5f415f4750363932676331745f6c5f5f6130376b5f5f674a35383675385a7843514c31525437525f3359394873373747525f374d5f5a5f4e5a6a53784c5a497a5f634776627766696356595f303337716c43475f315f345f5f415f305f63554858536a475f4845545f5f32635f47314d346b39615a486164325f5474755f5f315134466535665f615f7348596939396235305f525f5f49715f5f574d395f5f335f38333237555f4d5f684d5f585f72325f6651385f6a6d5f4e7a33386670533169355f71393831364b74704d765f337450435f5f32375f384c72756e30587a37764f365f35496c643032414a6833304c6e3901210100005f645f5f31625f4433445f45687078467a755f44305835635f386b33747745745638316438737436545f53656f5f36345f5662653658335f5f7146667a713938314f7355366c49355f635f65485f366d5f4d6d5f59333536746e505f774e5f35794d5f614e5f327078676d726e4241305f655f75306a5f3436713465437041584d33595f634a5f6d6e3156344a5434353546714938735f70413538355f6e485f67375f5f4c416f576f5a5f5f705f585f744f785f77485f6132546332585a36705f58375f5f5f3672666a5f3841673542556f645f5f76457147325f5136336934486e6d356f6b62585f63314d394e36685f5f324a38574b5f535a6c394f667541595f5f58465f5f5f5f6f48395f65775f415f385f6a6541364939325f79614170760304000000936d5e26e0f9f43e8a98697ed07e4556e5d940801d45f845420add744706de9c79a6319f225a0d92e01f15b5e673d787ab05194d7c39e126e2eafbc2fc1ccc6b3d907d242c8834e750cdf0f7f12797f1d8aced6448b14b7e6e491256539695ef57a99887cdd729ad768b12a694a37a0e524ec6d68b4a5bbd3c90956bb2d5360600011a8c4562b96c5cd31f713e4ce98a87e2ec8f3a340200000032001052ae5f5988c19190403f587a5c0b78d20f04baff26aa95285863225e59388a33ca29e356333e8b2bb269abcabea060c6d2096bc1d16328365503d3c9387500ec00cb6fef68ef30daef2f8cf757ad3701dcf33156f886a207b122e51e29ea5f20797b505cf78744d82f246ba08dbcd155725cfc13aa56dc1e796c60ef75c471b7040304000000936d5e26e0f9f43e8a98697ed07e4556e5d940801d45f845420add744706de9c79a6319f225a0d92e01f15b5e673d787ab05194d7c39e126e2eafbc2fc1ccc6b3d907d242c8834e750cdf0f7f12797f1d8aced6448b14b7e6e491256539695ef57a99887cdd729ad768b12a694a37a0e524ec6d68b4a5bbd3c90956bb2d536060101000000001c18f36818d40cd30339028abb2f9fb20f26621c77eadeeb07e61038367155a80100000000009ec477b416197d54553642028f81ca320e6f0f83281b51b408fd819e434b318b187405f0bf0fcfe5da31be12559eaed3a240b000974881fc7e14b21bd9227108',
+      'hex',
+    ),
     sectionHashes: {
-      0: Buffer.from('5b693f86a6a8053b79effacd031e2367a1d35cc64988795768920b2965013742', 'hex'),
-      1: Buffer.from('29e3fd2d0a8c786d5318be88f0be06629152ac26628396e28350f7c5b81b1d58', 'hex'),
-      2: Buffer.from('f09f9bf315fe3b244703f3695cafff63b67156f799dc5c0742d1612cdd4897be', 'hex'),
-      3: Buffer.from('d05ef8971a0464a35ab4d08efea23d6c9c86aeeb5e1e6992956ed814d0a3d761', 'hex'),
-      0xff: Buffer.from('c7fec5279e22792a9cad6346f8933c1b2249043e1a03c835030d4e71dfbac3e0', 'hex'),
+      0: Buffer.from('936d5e26e0f9f43e8a98697ed07e4556e5d940801d45f845420add744706de9c', 'hex'),
+      1: Buffer.from('79a6319f225a0d92e01f15b5e673d787ab05194d7c39e126e2eafbc2fc1ccc6b', 'hex'),
+      2: Buffer.from('3d907d242c8834e750cdf0f7f12797f1d8aced6448b14b7e6e491256539695ef', 'hex'),
+      3: Buffer.from('57a99887cdd729ad768b12a694a37a0e524ec6d68b4a5bbd3c90956bb2d53606', 'hex'),
+      4: Buffer.from('0f1a569cf5bb3688fc70cfa162b8d717b3aa88279d54c64fe50b2f84a8fc02d4', 'hex'),
+      5: Buffer.from('1122c460bebe621afb386ac2ab75b4fdc46ad521fb12b2cc3f2cbda11cb9993a', 'hex'),
+      0xff: Buffer.from('b38a61dd0f1ea2430905e13e3287353cede3bd3c2f37b5d40dd16ca1019cf650', 'hex'),
     } as { [index: number]: Buffer },
   },
 ]
@@ -144,21 +126,28 @@ describe.each(models)('Transactions', function (m) {
       expect(resp).toHaveProperty('signature')
 
       const signature = resp.signature ?? new Signature()
-      expect(signature.pubkey).toEqual(resp_addr.publicKey);
-
+      expect(signature.rawPubkey).toEqual(resp_addr.rawPubkey)
+      console.log(signature)
       // Verify raw signature
       const unsignedRawSigHash = hashSignatureSec([], signature.raw_salt, data.sectionHashes, signature.raw_indices, null, null)
-      const rawSig = ed25519.verify(signature.raw_signature.subarray(1), unsignedRawSigHash, signature.pubkey.subarray(1))
+      const rawSig = ed25519.verify(signature.raw_signature.subarray(1), unsignedRawSigHash, signature.rawPubkey.subarray(1))
 
       // Verify wrapper signature
-      const prefix = new Uint8Array([0x03]);
-      const rawHash: Buffer = hashSignatureSec([signature.pubkey], signature.raw_salt, data.sectionHashes, signature.raw_indices, signature.raw_signature, prefix);
-      const tmpHashes = {...data.sectionHashes};
+      const prefix = new Uint8Array([0x03])
+      const rawHash: Buffer = hashSignatureSec(
+        [signature.rawPubkey],
+        signature.raw_salt,
+        data.sectionHashes,
+        signature.raw_indices,
+        signature.raw_signature,
+        prefix,
+      )
+      const tmpHashes = { ...data.sectionHashes }
 
-      tmpHashes[Object.keys(tmpHashes).length - 1] = rawHash;
+      tmpHashes[Object.keys(tmpHashes).length - 1] = rawHash
 
-      const unsignedWrapperSigHash = hashSignatureSec([], signature.wrapper_salt, tmpHashes, signature.wrapper_indices, null, null);
-      const wrapperSig = ed25519.verify(signature.wrapper_signature.subarray(1), unsignedWrapperSigHash, resp_addr.publicKey.subarray(1));
+      const unsignedWrapperSigHash = hashSignatureSec([], signature.wrapper_salt, tmpHashes, signature.wrapper_indices, null, null)
+      const wrapperSig = ed25519.verify(signature.wrapper_signature.subarray(1), unsignedWrapperSigHash, resp_addr.rawPubkey.subarray(1))
 
       expect(wrapperSig && rawSig).toEqual(true)
     } finally {
