@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2018 - 2022 Zondax AG
+*   (c) 2018 - 2024 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -14,31 +14,192 @@
 *  limitations under the License.
 ********************************************************************************/
 #![no_std]
+#![no_main]
 #![no_builtins]
 #![allow(dead_code, unused_imports)]
 
-extern crate core;
-#[cfg(test)]
-#[macro_use]
-extern crate hex_literal;
-
-use core::convert::TryInto;
-use core::mem;
 use core::panic::PanicInfo;
-use core::slice::{from_raw_parts, from_raw_parts_mut};
 
-use curve25519_dalek::scalar::Scalar;
-use merlin::{Transcript, TranscriptRng, TranscriptRngBuilder};
-use rand::RngCore;
-use schnorrkel::context::{SigningContext, SigningTranscript};
-use schnorrkel::{PublicKey, SecretKey};
-use zeroize::Zeroize;
-
-use crate::bolos::*;
-
-mod bolos;
+use constants::{DIV_DEFAULT_LIST_LEN, DIV_SIZE, SPENDING_KEY_GENERATOR, KEY_DIVERSIFICATION_PERSONALIZATION, GH_FIRST_BLOCK};
+mod constants;
+use aes::Aes256;
+use aes::cipher::{
+    BlockCipher, BlockEncrypt, BlockDecrypt, NewBlockCipher,
+    generic_array::{GenericArray,typenum::U32},
+};
+use binary_ff1::BinaryFF1;
+use jubjub::{AffinePoint, ExtendedPoint, Fr};
 
 fn debug(_msg: &str) {}
+
+// ParserError should mirror parser_error_t from parser_common.
+// At the moment, just implement OK or Error
+#[repr(C)]
+pub enum ParserError {
+    ParserOk = 0,
+    ParserUnexpectedError = 5,
+}
+
+#[repr(C)]
+pub enum ConstantKey {
+    SpendingKeyGenerator,
+    ProofGenerationKeyGenerator,
+    ValueCommitmentRandomnessGenerator,
+}
+
+#[no_mangle]
+pub extern "C" fn from_bytes_wide(input: &[u8; 64], output: &mut [u8; 32]) -> ParserError {
+    let result = Fr::from_bytes_wide(input).to_bytes();
+    output.copy_from_slice(&result[0..32]);
+    ParserError::ParserOk
+}
+
+#[no_mangle]
+pub extern "C" fn scalar_multiplication(
+    input: &[u8; 32],
+    key: ConstantKey,
+    output: *mut [u8; 32],
+) -> ParserError {
+    let key_point = match key {
+        ConstantKey::SpendingKeyGenerator => constants::SPENDING_KEY_GENERATOR,
+        ConstantKey::ProofGenerationKeyGenerator => constants::PROOF_GENERATION_KEY_GENERATOR,
+        ConstantKey::ValueCommitmentRandomnessGenerator => constants::VALUE_COMMITMENT_RANDOMNESS_GENERATOR,
+    };
+
+    let extended_point = key_point.multiply_bits(input);
+    let result = AffinePoint::from(&extended_point);
+
+    unsafe {
+        let output_slice = &mut *output;
+        output_slice.copy_from_slice(&result.to_bytes());
+    }
+
+    ParserError::ParserOk
+}
+
+#[inline(never)]
+pub fn get_diversifiers(
+    dk: &[u8; 32],
+    start_index: &mut [u8; 11],
+    result: &mut [u8; 44],
+) {
+    // Initialize cipher
+    let key = GenericArray::from_slice(dk);
+    let cipher = Aes256::new(&key);
+    let mut scratch = [0; 12];
+
+    let mut ff1 = BinaryFF1::new(&cipher, 11, &[], &mut scratch).unwrap();
+
+    let mut d: [u8; 11];
+    let size = 4;
+
+    for c in 0..size {
+        d = *start_index;
+        ff1.encrypt(&mut d).unwrap();
+        result[c * 11..(c + 1) * 11].copy_from_slice(&d);
+        for k in 0..11 {
+            start_index[k] = start_index[k].wrapping_add(1);
+            if start_index[k] != 0 {
+                // No overflow
+                break;
+            }
+        }
+    }
+}
+
+#[no_mangle]
+pub extern "C" fn get_default_diversifier_list(
+    dk: &[u8; 32],
+    start_index: &mut [u8; 11],
+    d_l: &mut [u8; 44],
+) -> ParserError {
+    let start = &mut *start_index;
+    let diversifier =  &mut *d_l;
+    get_diversifiers(dk,  start, diversifier);
+    ParserError::ParserOk
+}
+
+#[no_mangle]
+pub extern "C" fn is_valid_diversifier(
+    hash: &[u8; 32],
+) -> bool {
+    let u = AffinePoint::from_bytes(*hash);
+    if u.is_some().unwrap_u8() == 1 {
+        let q = u.unwrap().mul_by_cofactor();
+        return q != ExtendedPoint::identity();
+    }
+
+    false
+}
+
+#[no_mangle]
+pub extern "C" fn get_pkd(
+    ivk_ptr: &[u8; 32],
+    h: &[u8; 32],
+    pk_d: &mut [u8; 32],
+) -> ParserError {
+
+    let affine = AffinePoint::from_bytes(*h).unwrap();
+    let extended = ExtendedPoint::from(affine);
+    let cofactor = extended.mul_by_cofactor();
+    let p = cofactor.to_niels().multiply_bits(ivk_ptr);
+    *pk_d = AffinePoint::from(p).to_bytes();
+
+    ParserError::ParserOk
+}
+
+#[no_mangle]
+pub extern "C" fn randomized_secret_from_seed(
+    ask:  &[u8; 32],
+    alpha:  &[u8; 32],
+    output:  &mut [u8; 32],
+) -> ParserError{
+
+    let mut skfr = Fr::from_bytes(ask).unwrap();
+    let alphafr = Fr::from_bytes(alpha).unwrap();
+    skfr += alphafr;
+    output.copy_from_slice(&skfr.to_bytes());
+
+    ParserError::ParserOk
+}
+
+#[no_mangle]
+pub extern "C" fn compute_sbar(
+    s:  &[u8; 32],
+    r:  &[u8; 32],
+    rsk:  &[u8; 32],
+    sbar:  &mut [u8; 32],
+) -> ParserError{
+    let s_point = Fr::from_bytes(s).unwrap();
+    let r_point = Fr::from_bytes(r).unwrap();
+    let rsk_point = Fr::from_bytes(rsk).unwrap();
+
+    let sbar_tmp = r_point + s_point * rsk_point;
+    sbar.copy_from_slice(&sbar_tmp.to_bytes());
+    ParserError::ParserOk
+}
+
+#[no_mangle]
+pub extern "C" fn add_points(
+    hash:  &[u8; 32],
+    value: &[u8; 32],
+    scalar:  &[u8; 32],
+    cv:  &mut [u8; 32]) -> ParserError{
+
+    let hash_point = AffinePoint::from_bytes(*hash).unwrap();
+    let hash_point_ex = ExtendedPoint::from(hash_point);
+    let cofactor = hash_point_ex.mul_by_cofactor();
+
+    let val = Fr::from_bytes(value).unwrap();
+    
+    let scale = AffinePoint::from_bytes(*scalar).unwrap();
+    let scale_extended = ExtendedPoint::from(scale);
+
+    let s = cofactor*val + scale_extended;
+    let vcm = AffinePoint::from(s).to_bytes();
+    cv.copy_from_slice(&vcm);
+    ParserError::ParserOk
+}
 
 #[cfg(not(test))]
 #[panic_handler]
@@ -46,236 +207,29 @@ fn panic(_info: &PanicInfo) -> ! {
     loop {}
 }
 
-#[inline(never)]
-fn mult_with_secret(k: &mut Scalar, sk: &[u8]) {
-    let mut skbytes = [0u8; 32];
-    skbytes.copy_from_slice(&sk[0..32]);
-    let s = Scalar::from_bits(skbytes);
-    *k *= s;
-}
-
-#[inline(never)]
-fn add_witness(k: &mut Scalar, x: [u8; 32]) -> [u8; 32] {
-    let r = Scalar::from_bits(x);
-    *k += r;
-    k.to_bytes()
-}
-
-#[inline(never)]
-fn get_challenge_scalar(k: &mut Scalar, tr: &mut Transcript) {
-    let mut kbytes = [0u8; 64];
-    tr.challenge_bytes(b"sign:c", &mut kbytes);
-    *k += Scalar::from_bytes_mod_order_wide(&kbytes);
-}
-
-#[inline(never)]
-fn get_witness_bytes_custom(br: &mut Transcript, nonce_seeds: &[&[u8]]) -> [u8; 32] {
-    c_zemu_log_stack(b"witness_bytes\x00".as_ref());
-    let mut x = [0u8; 32];
-    for ns in nonce_seeds {
-        br.append_message(b"nonce-bytes", ns);
-    }
-    {
-        let random_bytes = {
-            let mut bytes = [0u8; 32];
-            Trng.fill_bytes(&mut bytes);
-            bytes
-        };
-        br.append_message(b"rng", &random_bytes);
-    }
-    br.challenge_bytes(b"witness-bytes", &mut x);
-    br.zeroize();
-    x
-}
-
-#[no_mangle]
-pub extern "C" fn sign_sr25519_phase1(
-    sk_ristretto_expanded_ptr: *const u8,
-    pk_ptr: *const u8,
-    context_ptr: *const u8,
-    context_len: usize,
-    msg_ptr: *const u8,
-    msg_len: usize,
-    sig_ptr: *mut u8,
-) {
-    c_zemu_log_stack(b"sign_sr25519\x00".as_ref());
-
-    let sk_ristretto_expanded =
-        unsafe { from_raw_parts(sk_ristretto_expanded_ptr as *const u8, 64) };
-    let pk = unsafe { from_raw_parts(pk_ptr as *const u8, 32) };
-    let context = unsafe { from_raw_parts(context_ptr as *const u8, context_len) };
-    let message = unsafe { from_raw_parts(msg_ptr as *const u8, msg_len) };
-    let signature = unsafe { from_raw_parts_mut(sig_ptr as *mut u8, 64) };
-
-    let mut signtranscript = Transcript::new(b"SigningContext");
-    signtranscript.append_message(b"", context);
-    signtranscript.append_message(b"sign-bytes", message);
-    signtranscript.append_message(b"proto-name", b"Schnorr-sig"); //proto name
-    signtranscript.append_message(b"sign:pk", pk); //commitpoint: pk
-
-    let x = get_witness_bytes_custom(&mut signtranscript, &[&sk_ristretto_expanded[32..]]);
-    signature[32..64].copy_from_slice(&x);
-}
-
-#[no_mangle]
-pub extern "C" fn sign_sr25519_phase2(
-    sk_ristretto_expanded_ptr: *const u8,
-    pk_ptr: *const u8,
-    context_ptr: *const u8,
-    context_len: usize,
-    msg_ptr: *const u8,
-    msg_len: usize,
-    sig_ptr: *mut u8,
-) {
-    c_zemu_log_stack(b"sign_sr25519\x00".as_ref());
-
-    let sk_ristretto_expanded =
-        unsafe { from_raw_parts(sk_ristretto_expanded_ptr as *const u8, 64) };
-    let pk = unsafe { from_raw_parts(pk_ptr as *const u8, 32) };
-    let context = unsafe { from_raw_parts(context_ptr as *const u8, context_len) };
-    let message = unsafe { from_raw_parts(msg_ptr as *const u8, msg_len) };
-    let signature = unsafe { from_raw_parts_mut(sig_ptr as *mut u8, 64) };
-
-    let mut signtranscript = Transcript::new(b"SigningContext");
-    signtranscript.append_message(b"", context);
-    signtranscript.append_message(b"sign-bytes", message);
-    signtranscript.append_message(b"proto-name", b"Schnorr-sig"); //proto name
-    signtranscript.append_message(b"sign:pk", pk); //commitpoint: pk
-    signtranscript.append_message(b"sign:R", &signature[0..32]); //commitpoint: pk
-
-    let mut x = [0u8; 32];
-    x.copy_from_slice(&signature[32..64]);
-
-    let mut k = Scalar::zero();
-    get_challenge_scalar(&mut k, &mut signtranscript);
-
-    mult_with_secret(&mut k, sk_ristretto_expanded);
-    signature[32..].copy_from_slice(&add_witness(&mut k, x));
-    signature[63] |= 128;
-}
-
-#[no_mangle]
-pub extern "C" fn get_sr25519_sk(sk_ed25519_expanded_ptr: *mut u8) {
-    let sk_ed25519_expanded = unsafe { from_raw_parts_mut(sk_ed25519_expanded_ptr as *mut u8, 64) };
-    let secret: SecretKey = SecretKey::from_ed25519_bytes(&sk_ed25519_expanded[..]).unwrap();
-    sk_ed25519_expanded.copy_from_slice(&secret.to_bytes());
-}
-
 #[cfg(test)]
 mod tests {
-    use curve25519_dalek::constants::RISTRETTO_BASEPOINT_POINT;
-    use curve25519_dalek::edwards::EdwardsPoint;
-    use curve25519_dalek::scalar::Scalar;
-    use log::{debug, info};
-    use schnorrkel::{context::*, Keypair, PublicKey, SecretKey, Signature};
+    // use curve25519_dalek::constants::RISTRETTO_BASEPOINT_POINT;
+    // use curve25519_dalek::edwards::EdwardsPoint;
+    // use curve25519_dalek::scalar::Scalar;
+    // use log::{debug, info};
+    // use schnorrkel::{context::*, Keypair, PublicKey, SecretKey, Signature};
 
-    use crate::*;
-    use core::ops::Mul;
+    // use crate::*;
+    // use core::ops::Mul;
 
-    fn init_logging() {
-        let _ = env_logger::builder().is_test(true).try_init();
-    }
+    // fn init_logging() {
+    //     let _ = env_logger::builder().is_test(true).try_init();
+    // }
 
-    fn ristretto_scalarmult(sk: &[u8], pk: &mut [u8]) {
-        let mut seckey = [0u8; 32];
-        seckey.copy_from_slice(&sk[0..32]);
-        let pubkey = RISTRETTO_BASEPOINT_POINT
-            .mul(Scalar::from_bits(seckey))
-            .compress()
-            .0;
-        pk.copy_from_slice(&pubkey);
-    }
+    // fn ristretto_scalarmult(sk: &[u8], pk: &mut [u8]) {
+    //     let mut seckey = [0u8; 32];
+    //     seckey.copy_from_slice(&sk[0..32]);
+    //     let pubkey = RISTRETTO_BASEPOINT_POINT
+    //         .mul(Scalar::from_bits(seckey))
+    //         .compress()
+    //         .0;
+    //     pk.copy_from_slice(&pubkey);
+    // }
 
-    #[test]
-    fn test_sign_verify() {
-        let mut sk_ed25519_expanded = [
-            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05,
-            0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03,
-            0x04, 0x05, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x00, 0x01,
-            0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
-            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
-        ];
-
-        let pk_expected = "b65abc66a8fdeac1197d03daa6c3791d0c0799a52db6b7127b1cd12d46e34364";
-
-        let secret = SecretKey::from_ed25519_bytes(&sk_ed25519_expanded).unwrap();
-
-        let mut pk = [0u8; 32];
-        get_sr25519_sk(sk_ed25519_expanded.as_mut_ptr());
-
-        ristretto_scalarmult(&sk_ed25519_expanded, &mut pk);
-
-        assert_eq!(hex::encode(pk), pk_expected);
-
-        let context = b"good";
-        let msg = b"test message";
-        let mut signature = [0u8; 64];
-
-        sign_sr25519_phase1(
-            secret.to_bytes().as_ptr(),
-            pk.as_ptr(),
-            context.as_ptr(),
-            context.len(),
-            msg.as_ptr(),
-            msg.len(),
-            signature.as_mut_ptr(),
-        );
-
-        let mut x = [0u8; 32];
-        x.copy_from_slice(&signature[32..64]);
-
-        ristretto_scalarmult(&x, &mut signature[0..32]);
-
-        sign_sr25519_phase2(
-            secret.to_bytes().as_ptr(),
-            pk.as_ptr(),
-            context.as_ptr(),
-            context.len(),
-            msg.as_ptr(),
-            msg.len(),
-            signature.as_mut_ptr(),
-        );
-
-        let keypair: Keypair = Keypair::from(secret);
-
-        let mut sigledger = [0u8; 64];
-        hex::decode_to_slice("48fdbe5cf3524bdd078ac711565d658a3053d10660749959883c4710f49d9948b2d5f829bea6800897dc6ea0150ca11075cc36b75bfcf3712aafb8e1bd10bf8f",&mut sigledger).expect("dec");
-
-        let self_sig = Signature::from_bytes(&signature).unwrap();
-        let self_sig_ledger = Signature::from_bytes(&sigledger).unwrap();
-
-        let vers = signing_context(context);
-
-        assert!(
-            keypair.verify(vers.bytes(msg), &self_sig).is_ok(),
-            "Verification of a valid signature failed!"
-        );
-        assert!(
-            keypair.verify(vers.bytes(msg), &self_sig_ledger).is_ok(),
-            "Verification of a valid signature from ledger failed!"
-        );
-    }
-
-    #[test]
-    fn get_public_key_c() {
-        init_logging();
-
-        let mut sk_ed25519_expanded = [
-            0x00, 0x01, 0x02, 0x03, 04, 0x5, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 04, 0x5, 0x06,
-            0x07, 0x00, 0x01, 0x02, 0x03, 04, 0x5, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 04, 0x5,
-            0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 04, 0x5, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 04,
-            0x5, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03, 04, 0x5, 0x06, 0x07, 0x00, 0x01, 0x02, 0x03,
-            04, 0x5, 0x06, 0x07,
-        ];
-
-        let pk_expected = "b65abc66a8fdeac1197d03daa6c3791d0c0799a52db6b7127b1cd12d46e34364";
-
-        let mut pk = [0u8; 32];
-        get_sr25519_sk(sk_ed25519_expanded.as_mut_ptr());
-
-        ristretto_scalarmult(&sk_ed25519_expanded, &mut pk);
-
-        info!("{:?}", hex::encode(pk));
-        assert_eq!(hex::encode(pk), pk_expected);
-    }
 }
