@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2018 - 2022 Zondax AG
+*   (c) 2018 - 2024 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -16,4 +16,4 @@
 
 #include "actions.h"
 
-address_state_t action_addrResponse;
+uint16_t cmdResponseLen;
