@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2018 - 2022 Zondax AG
+*   (c) 2018 - 2024 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -25,32 +25,75 @@
 #include "zxerror.h"
 #include "parser_txdef.h"
 #include "zxformat.h"
+#include "nvdata.h"
 
-typedef struct {
-    address_kind_e kind;
-    uint16_t len;
-} address_state_t;
-
-typedef struct {
-    key_type_e kind;
-    uint16_t len;
-} key_state_t;
-
-extern address_state_t action_addrResponse;
-extern key_state_t key_state;
+extern uint16_t cmdResponseLen;
 
 __Z_INLINE zxerr_t app_fill_address(signing_key_type_e addressKind) {
     // Put data directly in the apdu buffer
     zemu_log("app_fill_address\n");
     MEMZERO(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE);
 
-    action_addrResponse.len = 0;
+    cmdResponseLen = 0;
     zxerr_t err = crypto_fillAddress(addressKind,
                                      G_io_apdu_buffer, IO_APDU_BUFFER_SIZE,
-                                     &action_addrResponse.len);
+                                     &cmdResponseLen);
 
-    if (err != zxerr_ok || action_addrResponse.len == 0) {
+    if (err != zxerr_ok || cmdResponseLen == 0) {
         THROW(APDU_CODE_EXECUTION_ERROR);
+    }
+
+    return err;
+}
+
+__Z_INLINE zxerr_t app_fill_keys(key_kind_e requestedKey) {
+    // Put data directly in the apdu buffer
+    zemu_log("app_fill_keys\n");
+    MEMZERO(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE);
+
+    cmdResponseLen = 0;
+    zxerr_t err = crypto_fillMASP(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE,
+                                     &cmdResponseLen, requestedKey);
+
+    if (err != zxerr_ok || cmdResponseLen == 0) {
+        THROW(APDU_CODE_EXECUTION_ERROR);
+    }
+
+    return err;
+}
+
+__Z_INLINE zxerr_t app_fill_randomness(masp_type_e type) {
+    // Put data directly in the apdu buffer
+    zemu_log("app_fill_randomness\n");
+    MEMZERO(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE);
+
+    if (get_state() != STATE_INITIAL && get_state() != STATE_PROCESSED_RANDOMNESS) {
+        return zxerr_unknown;
+    }
+
+    cmdResponseLen = 0;
+    zxerr_t err = crypto_computeRandomness(type, G_io_apdu_buffer, IO_APDU_BUFFER_SIZE - 3, &cmdResponseLen);
+
+    if (err != zxerr_ok || cmdResponseLen == 0) {
+        transaction_reset();
+        THROW(APDU_CODE_DATA_INVALID);
+    }
+
+    set_state(STATE_PROCESSED_RANDOMNESS);
+    return err;
+}
+
+__Z_INLINE zxerr_t app_fill_spend_sig() {
+    // Put data directly in the apdu buffer
+    zemu_log("app_fill_spend_sig\n");
+    MEMZERO(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE);
+
+    cmdResponseLen = 0;
+    zxerr_t err = crypto_extract_spend_signature(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE - 3, &cmdResponseLen);
+
+    if (err != zxerr_ok || cmdResponseLen == 0) {
+        transaction_reset();
+        THROW(APDU_CODE_DATA_INVALID);
     }
 
     return err;
@@ -58,28 +101,47 @@ __Z_INLINE zxerr_t app_fill_address(signing_key_type_e addressKind) {
 
 __Z_INLINE void app_sign() {
     const parser_tx_t *txObj = tx_get_txObject();
+
     const zxerr_t err = crypto_sign(txObj, G_io_apdu_buffer, sizeof(G_io_apdu_buffer) - 2);
 
     if (err != zxerr_ok) {
+        transaction_reset();
         MEMZERO(G_io_apdu_buffer, sizeof(G_io_apdu_buffer));
         set_code(G_io_apdu_buffer, 0, APDU_CODE_SIGN_VERIFY_ERROR);
         io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, 2);
     } else {
+        transaction_reset();
         const uint16_t responseLen = PK_LEN_25519_PLUS_TAG + 2 * SALT_LEN + 2 * SIG_LEN_25519_PLUS_TAG + 2 + 10;
         set_code(G_io_apdu_buffer, responseLen, APDU_CODE_OK);
         io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, responseLen + 2);
     }
 }
 
+__Z_INLINE void app_sign_masp_spends() {
+    parser_tx_t *txObj = tx_get_txObject();
+    const zxerr_t err = crypto_sign_masp_spends(txObj, G_io_apdu_buffer, IO_APDU_BUFFER_SIZE - 3);
+
+    if (err != zxerr_ok) {
+        transaction_reset();
+        MEMZERO(G_io_apdu_buffer, sizeof(G_io_apdu_buffer));
+        set_code(G_io_apdu_buffer, 0, APDU_CODE_SIGN_VERIFY_ERROR);
+        io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, 2);
+    } else {
+        set_code(G_io_apdu_buffer, HASH_LEN, APDU_CODE_OK);
+        io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, HASH_LEN + 2);
+    }
+}
+
 __Z_INLINE void app_reject() {
+    transaction_reset();
     MEMZERO(G_io_apdu_buffer, IO_APDU_BUFFER_SIZE);
     set_code(G_io_apdu_buffer, 0, APDU_CODE_COMMAND_NOT_ALLOWED);
     io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, 2);
 }
 
-__Z_INLINE void app_reply_address() {
-    set_code(G_io_apdu_buffer, action_addrResponse.len, APDU_CODE_OK);
-    io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, action_addrResponse.len + 2);
+__Z_INLINE void app_reply_cmd() {
+    set_code(G_io_apdu_buffer, cmdResponseLen, APDU_CODE_OK);
+    io_exchange(CHANNEL_APDU | IO_RETURN_AFTER_TX, cmdResponseLen + 2);
 }
 
 __Z_INLINE void app_reply_error() {
