@@ -1,5 +1,5 @@
 /*******************************************************************************
-*  (c) 2018 - 2023 Zondax AG
+*  (c) 2018 - 2024 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -16,8 +16,15 @@
 #include "parser_impl_common.h"
 #include "leb128.h"
 
-#define SIGN_MASK 0x80000000
-#define SCALE_SHIFT 16
+bool isAllZeroes(const void *buf, size_t n) {
+    uint8_t *p = (uint8_t *) buf;
+    for (size_t i = 0; i < n; ++i) {
+        if (p[i]) {
+            return false;
+        }
+    }
+    return true;
+}
 
 parser_error_t readByte(parser_context_t *ctx, uint8_t *byte) {
     if (byte == NULL || ctx->offset >= ctx->bufferLen) {
@@ -59,67 +66,22 @@ parser_error_t readUint64(parser_context_t *ctx, uint64_t *value) {
     return parser_ok;
 }
 
-parser_error_t readUint256(parser_context_t *ctx, uint256_t *value) {
-    if (value == NULL || ctx->offset + sizeof(uint256_t) > ctx->bufferLen) {
-        return parser_unexpected_error;
-    }
-
-    MEMCPY(value, ctx->buffer + ctx->offset, sizeof(uint256_t));
-    ctx->offset += sizeof(uint256_t);
-    return parser_ok;
-}
-
-zxerr_t recover_decimal(const uint8_t* bytes, int64_t* num, uint32_t* scale) {
-    if (bytes == NULL) {
-        return zxerr_unknown; // Invalid byte sequence
-    }
-
-    uint32_t flag = 0;
-    for (int i = 0; i < 4; i++) {
-        flag |= ((uint32_t)bytes[i]) << (8 * i);
-    }
-
-    *scale = (flag >> SCALE_SHIFT);
-    uint8_t is_negative = (flag & SIGN_MASK) != 0;
-
-    uint32_t hi = 0, lo = 0, mid = 0;
-    for (int i = 0; i < 4; i++) {
-        hi |= ((uint32_t)bytes[4 + i]) << (8 * i);
-        lo |= ((uint32_t)bytes[8 + i]) << (8 * i);
-        mid |= ((uint32_t)bytes[12 + i]) << (8 * i);
-    }
-
-    uint64_t m = ((uint64_t)hi) << 32 | lo;
-    m |= ((uint64_t)mid) << 32;
-
-    if (is_negative) {
-        m = ~m + 1; // Two's complement negation
-    }
-
-    *num = (int64_t)m;
-    return zxerr_ok;
-}
-
-parser_error_t readDecimal(parser_context_t *ctx, serialized_decimal *value) {
-    if (value == NULL || ctx->offset + sizeof(serialized_decimal) > ctx->bufferLen) {
-        return parser_unexpected_error;
-    }
-    uint8_t raw_decimal[sizeof(serialized_decimal)] = {0};
-    MEMCPY(raw_decimal, ctx->buffer + ctx->offset, sizeof(serialized_decimal));
-
-    recover_decimal((const uint8_t *) &raw_decimal, &value->num, &value->scale);
-
-    ctx->offset += sizeof(serialized_decimal);
-    return parser_ok;
-}
-
-
 parser_error_t readBytes(parser_context_t *ctx, const uint8_t **output, uint16_t outputLen) {
     if (ctx->offset + outputLen > ctx->bufferLen) {
         return parser_unexpected_buffer_end;
     }
 
     *output = ctx->buffer + ctx->offset;
+    ctx->offset += outputLen;
+    return parser_ok;
+}
+
+parser_error_t readBytesSize(parser_context_t *ctx, uint8_t *output, uint16_t outputLen) {
+    if (ctx->offset + outputLen > ctx->bufferLen) {
+        return parser_unexpected_buffer_end;
+    }
+
+    MEMCPY(output, (ctx->buffer + ctx->offset), outputLen);
     ctx->offset += outputLen;
     return parser_ok;
 }
@@ -139,6 +101,44 @@ parser_error_t readFieldSize(parser_context_t *ctx, uint32_t *size) {
     return parser_ok;
 }
 
+parser_error_t readFieldSizeU16(parser_context_t *ctx, uint16_t *size) {
+    uint8_t consumed = 0;
+    uint64_t tmpSize = 0;
+
+    decodeLEB128(ctx->buffer + ctx->offset, 10, &consumed, &tmpSize);
+    ctx->offset += consumed;
+
+    if (tmpSize > UINT16_MAX) {
+        return parser_value_out_of_range;
+    }
+    *size = (uint16_t)tmpSize;
+
+    return parser_ok;
+}
+
+parser_error_t readCompactSize(parser_context_t *ctx, uint64_t *result) {
+    uint8_t tag = 0;
+    uint16_t tmp16 = 0;
+    uint32_t tmp32 = 0;
+    CHECK_ERROR(readByte(ctx, &tag))
+    switch(tag) {
+    case 253:
+        CHECK_ERROR(readUint16(ctx, &tmp16))
+        *result = (uint64_t)tmp16;
+        break;
+    case 254:
+        CHECK_ERROR(readUint32(ctx, &tmp32))
+        *result = (uint64_t)tmp32;
+        break;
+    case 255:
+        CHECK_ERROR(readUint64(ctx, result))
+        break;
+    default:
+        *result = (uint64_t)tag;
+    }
+    return parser_ok;
+}
+
 parser_error_t checkTag(parser_context_t *ctx, uint8_t expectedTag) {
     uint8_t tmpTag = 0;
     CHECK_ERROR(readByte(ctx, &tmpTag))
@@ -146,5 +146,21 @@ parser_error_t checkTag(parser_context_t *ctx, uint8_t expectedTag) {
     if (tmpTag != expectedTag) {
         return parser_unexpected_value;
     }
+    return parser_ok;
+}
+
+parser_error_t readPubkey(parser_context_t *ctx, bytes_t *pubkey) {
+    if (ctx == NULL || pubkey == NULL) {
+        return parser_unexpected_buffer_end;
+    }
+
+    if (ctx->offset >= ctx->bufferLen) {
+        return parser_unexpected_buffer_end;
+    }
+
+    const uint8_t pkType = *(ctx->buffer + ctx->offset);
+    //Pubkey must include pkType (needed for encoding)
+    pubkey->len = 1 + (pkType == key_ed25519 ? PK_LEN_25519 : COMPRESSED_SECP256K1_PK_LEN);
+    CHECK_ERROR(readBytes(ctx, &pubkey->ptr, pubkey->len))
     return parser_ok;
 }
