@@ -1,5 +1,5 @@
 /*******************************************************************************
-*  (c) 2018 - 2022 Zondax AG
+*  (c) 2018 - 2024 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -30,22 +30,27 @@ parser_error_t printTxnFields(const parser_context_t *ctx,
                               char *outVal, uint16_t outValLen,
                               uint8_t pageIdx, uint8_t *pageCount);
 
+parser_error_t printMemo( const parser_context_t *ctx,
+                        char *outKey, uint16_t outKeyLen,
+                        char *outVal, uint16_t outValLen,
+                        uint8_t pageIdx, uint8_t *pageCount);
+
 parser_error_t printExpert(const parser_context_t *ctx,
                            uint8_t displayIdx,
                            char *outKey, uint16_t outKeyLen,
                            char *outVal, uint16_t outValLen,
                            uint8_t pageIdx, uint8_t *pageCount);
 
-parser_error_t printCodeHash(bytes_t *codeHash,
+parser_error_t printCodeHash(section_t *codeSection,
                              char *outKey, uint16_t outKeyLen,
                              char *outVal, uint16_t outValLen,
                              uint8_t pageIdx, uint8_t *pageCount);
 
-parser_error_t printAddress(bytes_t pubkeyHash,
+parser_error_t printAddressAlt(const AddressAlt *addr,
                             char *outVal, uint16_t outValLen,
                             uint8_t pageIdx, uint8_t *pageCount);
 
-parser_error_t printAmount( const uint256_t *amount, uint8_t amountDenom, const char* symbol,
+parser_error_t printAmount( const bytes_t *amount, bool isSigned, uint8_t amountDenom, const char* symbol,
                             char *outVal, uint16_t outValLen,
                             uint8_t pageIdx, uint8_t *pageCount);
 
@@ -53,7 +58,13 @@ parser_error_t printPublicKey(const bytes_t *pubkey,
                               char *outVal, uint16_t outValLen,
                               uint8_t pageIdx, uint8_t *pageCount);
 
-parser_error_t uint256_to_str(char *output, uint16_t outputLen, const uint256_t *value);
+parser_error_t joinStrings(const bytes_t first, const bytes_t second, const char *separator,
+                            char * outVal, uint16_t outValLen, uint8_t pageIdx, uint8_t *pageCount);
+
+parser_error_t printProposal( const tx_init_proposal_t *initProposal, uint8_t displayIdx,
+                                   char *outKey, uint16_t outKeyLen,
+                                   char *outVal, uint16_t outValLen,
+                                   uint8_t pageIdx, uint8_t *pageCount);
 
 #ifdef __cplusplus
 }
